# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AirFlex API",
    "version": "1.0.0",
    "description": "Open-source P2P Airtime & Data Exchange Marketplace backed by the Stellar Network. All protected endpoints require a Bearer JWT obtained from POST /api/v1/auth/verify-otp.",
    "contact": {
      "name": "AirFlex Core Team",
      "url": "https://github.com/arflexx/Airflex"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/v1/auth/verify-otp. Pass as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Descriptive error message"
          }
        }
      },
...
      "TradeOffer": {
        "type": "object",
        "required": [
          "id",
          "asset_type",
          "amount",
          "status",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Seller UUID. Present on detail/create responses but omitted from the public listing feed, which exposes `seller_handle` instead (issue #330)."
          },
          "seller_handle": {
            "type": "string",
            "nullable": true,
            "example": "@airflex_2f9a",
            "description": "Opaque, deterministic seller label derived from the user id. Included on the public listing feed instead of `seller_id` (issue #330)."
          },
          "buyer_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "nullable": true
          },
          "asset_type": {
            "type": "string",
            "example": "MTN_AIRTIME"
          },
          "amount": {
            "type": "number",
            "example": 500
          },
          "status": {
            "$ref": "#/components/schemas/TradeStatus"
          },
          "contract_listing_id": {
            "type": [
              "string",
              "null"
            ],
            "nullable": true,
            "example": "42"
          },
          "escrow_tx_hash": {
            "type": [
              "string",
              "null"
            ],
            "nullable": true,
            "example": "abc123txhash"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTradeBody": {
        "type": "object",
        "required": [
          "assetType",
          "amount",
          "expiresInHours"
        ],
        "properties": {
          "assetType": {
            "type": "string",
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9_-]+$",
            "example": "MTN_AIRTIME"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0,
            "example": 500
          },
          "expiresInHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 168,
            "example": 24
          }
        }
      },
...
      "TradeListResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TradeOffer"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
      "DeleteAccountResponse": {
        "type": "object",
        "required": [
          "message",
          "scheduledDeletionAt"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Account deletion requested. Your account will be anonymised in 30 days."
          },
          "scheduledDeletionAt": {
            "type": "string",
            "format": "date-time",
            "description": "ISO timestamp 30 days from now."
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid Bearer JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Missing or invalid Authorization header"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Trade offer not found"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Request body failed schema validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/trades": {
      "get": {
        "tags": [
          "Trades"
        ],
        "summary": "List active trade offers",
        "description": "Returns paginated active listings that have not expired.",
        "security": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of active trade offers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "Trades"
        ],
        "summary": "Create a trade offer",
        "description": "Creates a new trade listing. Calls `create_listing` on the Soroban escrow contract and stores the resulting on-chain listing ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTradeBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trade offer created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TradeOffer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Seller wallet not found or contract call failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get own profile",
        "description": "Returns the authenticated user's profile metadata. Phone is masked.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "summary": "Request account deletion (GDPR/NDPR)",
        "description": "Initiates a 30-day soft-delete grace period. The account is marked `pending_deletion`. After 30 days all PII is anonymised. All pending trades are cancelled and escrow refunds are triggered. A confirmation SMS is sent via Termii.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion request accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Account deletion already pending.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/profile/trades": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get own trade history",
        "description": "Returns paginated trade history for the authenticated user (as seller or buyer). Filterable by status.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "All",
                "Active",
                "Locked",
                "Completed",
                "Cancelled",
                "Disputed"
              ],
              "default": "All"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated trade history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileTradesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/profile/deletion-status": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Check account deletion status",
        "description": "Returns whether the account has a pending deletion request and when it is scheduled.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/profile/cancel-deletion": {
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "Cancel pending account deletion",
        "description": "Cancels a pending 30-day deletion request within the grace window.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion request cancelled successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Account deletion cancelled. Your account is restored."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No pending deletion request to cancel.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
}